Spatial: Add rotateAround() method, replaces arcballRotate() from ArcballCameraControl
