+++ src/main/kotlin/graphics/scenery/controls/behaviours/ArcballCameraControl.kt
@@ -114,34 +114,12 @@ open class ArcballCameraControl(private val name: String, camera: () -> Camera?,
             lastX = x
             lastY = y
 
-            arcballRotate(xoffset, yoffset, node)
+            node.spatial().rotateAround(xoffset, yoffset, target.invoke())
 
             node.lock.unlock()
         }
     }
 
-    /**
-     *  This function rotates the [node] by a fixed [yaw] and [pitch] about the [target]
-     *
-     *  @param[yaw] yaw in degrees
-     *  @param[pitch] pitch in degrees
-     *  @param[node] the camera to move
-     */
-    private fun arcballRotate(yaw: Float, pitch: Float, node: Camera) {
-        val frameYaw = (yaw) / 180.0f * Math.PI.toFloat()
-        val framePitch = pitch / 180.0f * Math.PI.toFloat()
-
-        // first calculate the total rotation quaternion to be applied to the camera
-        val yawQ = Quaternionf().rotateXYZ(0.0f, frameYaw, 0.0f).normalize()
-        val pitchQ = Quaternionf().rotateXYZ(framePitch, 0.0f, 0.0f).normalize()
-
-        node.ifSpatial {
-            distance = (target.invoke() - position).length()
-            node.target = target.invoke()
-            rotation = pitchQ.mul(rotation).mul(yawQ).normalize()
-            position = target.invoke() + node.forward * distance * (-1.0f)
-        }
-    }
 
     /**
      *  This function rotates the camera controlled by this behaviour by a fixed [yaw]
@@ -156,7 +134,7 @@ open class ArcballCameraControl(private val name: String, camera: () -> Camera?,
                 return
             }
 
-            arcballRotate(yaw, pitch, node)
+            node.spatial().rotateAround(yaw, pitch, target.invoke())
 
             node.lock.unlock()
         }
